requirements.txt
flask==3.0.3
tensorflow==2.16.1
pillow==10.3.0
numpy==1.26.4
gunicorn==22.0.0
scikit-learn==1.4.2
matplotlib==3.8.4